% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundEntropy.R
\name{boundEntropy}
\alias{boundEntropy}
\title{boundEntropy function}
\usage{
boundEntropy(H, n, totalCharacters = 21)
}
\arguments{
\item{H}{The is the calculated H, entropy value, to be bounded by the maximum entropy
and minimum entropy possible.}

\item{n}{By default, n = 1, which means only a single character per alignment
column per sequence. For example, users may change to n = 9 for nonamer 
scoring instead of monomer.}

\item{totalCharacters}{This depends on whether the standard 20 amino acids plus
a gap scoring is used or not. Users may set this to 20 if they wish to ignore
gap character.}
}
\description{
This function is one of the functions that makes EntropyMSA works.
}
\examples{
#see EntropyMSA function
}
\keyword{entropy}

