% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_vect.R
\name{entropy_vect}
\alias{entropy_vect}
\title{entropy_vect function}
\usage{
entropy_vect(vect)
}
\arguments{
\item{vect}{The input for entropy calculation in vector form.}
}
\description{
This function is one of the functions that makes EntropyMSA works.
}
\examples{
#see EntropyMSA function
}
\keyword{entropy}

