% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta2list.R
\name{fasta2list}
\alias{fasta2list}
\title{fasta2list function}
\usage{
fasta2list(MSAfile)
}
\arguments{
\item{MSAfile}{The input is a multiple sequence alignment in FASTA format.}
}
\description{
This function is one of the functions that makes EntropyMSA works.
}
\examples{
#see EntropyMSA function
}
\keyword{entropy}

