% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_removal.R
\name{gap_removal}
\alias{gap_removal}
\title{gap_removal function}
\usage{
gap_removal(MSA2, gap)
}
\arguments{
\item{MSA2}{The input is a multiple sequence alignment in FASTA format.}

\item{gap}{This is for gap removal in an alignment column. The default 
accepts 100 percentage of gaps. Users can change the numeric value provided to
this parameter from 0 to 1 to change the gaps that can be tolerated per
alignment column. 0 for 0 percentage gaps, 1 for 100 percentage gaps.}
}
\description{
This function is one of the functions that makes EntropyMSA works.
}
\examples{
#see EntropyMSA function
}
\keyword{entropy}

