% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_gap_firstseq.R
\name{no_gap_firstseq}
\alias{no_gap_firstseq}
\title{no_gap_firstseq function}
\usage{
no_gap_firstseq(MSAfile, no.gap.firstseq)
}
\arguments{
\item{MSAfile}{The input is a multiple sequence alignment in FASTA format.}

\item{no.gap.firstseq}{This allows the user to set gap removal based on a
reference sequence that must be placed as the first sequence in the
multiple alignment file.}
}
\description{
This function is one of the functions that makes EntropyMSA works.
}
\examples{
#see EntropyMSA function
}
\keyword{entropy}

