% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampler.R
\name{resampler}
\alias{resampler}
\title{resampler function}
\usage{
resampler(vect, randomise)
}
\arguments{
\item{vect}{The input for entropy calculation in vector form.}

\item{randomise}{This parameter allows users to try other corrected entropy
values. By default, set.seed is used in entropy correction and hence, same
values are produced, which made it easier for the empiral study to determine
the effects of correction in various sample and population sizes.}
}
\description{
This function is one of the functions that makes EntropyMSA works.
}
\examples{
#see EntropyMSA function
}
\keyword{entropy}

