% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_size.R
\name{window_size}
\alias{window_size}
\title{window_size function}
\usage{
window_size(MSA3, n)
}
\arguments{
\item{MSA3}{The input is a multiple sequence alignment in FASTA format.}

\item{n}{By default, n = 1, which means only a single character per alignment
column per sequence. For example, users may change to n = 9 for nonamer 
scoring instead of monomer.}
}
\description{
This function is one of the functions that makes EntropyMSA works.
}
\examples{
#see EntropyMSA function
}
\keyword{entropy}

