% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writealn.R
\name{writealn}
\alias{writealn}
\title{writealn function}
\usage{
writealn(MSA3, filename)
}
\arguments{
\item{MSA3}{The input is a multiple sequence alignment in FASTA format.}

\item{filename}{This is the filename of the output alignment.}
}
\description{
This function is one of the functions that makes EntropyMSA works.
}
\examples{
#see EntropyMSA function
}
\keyword{entropy}

